JAM: small fix due to compiler problem gcc4.5.2/64 bit
git-svn-id: https://subversion.gsi.de/go4/trunk@861 302e7798-3865-0410-85c9-d38765bbeff6

+++ applications/hadaq-go4/THadaqUnpackProc.cxx
@@ -753,7 +753,7 @@ THadaqUnpackProc::BuildEvent(TGo4EventElement* dest)
         }
       char* cursor = (char*) pdata;
       char* datastart = (char*) pdata;	
-       EPRINT("*** Start cursor #0x%x\n",  (unsigned) cursor);
+       EPRINT("*** Start cursor #0x%x\n",  (unsigned long) cursor);
 	
 
       Int_t bufbytes=lwords*sizeof(Int_t);
